Extract multiple classes in different modules
Most classes that were in connect.py were moved to other modules.
I tried to eliminate circular dependencies, that's why there is a
CommandFactory and a CommandFactoryImplementation. I couldn't think of a
more appropriate name, so it shall remain so for now.

I think the addition of the CommandId class was a nice touch.